@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.5.2</version>
+		<version>3.0.1</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>ar.latorraca</groupId>
@@ -14,7 +14,7 @@
 	<name>personal-finance-api</name>
 	<description>Personal Finance API</description>
 	<properties>
-		<java.version>16</java.version>
+		<java.version>19</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
